Restructured Rent-Util and all files that require the module. Removed remaining unnecessary ComponentScans.
