+++ Rent/Rent-API/src/main/java/com/rent/api/configuration/OAuth2ResourceServerConfig.java
@@ -2,7 +2,6 @@ package com.rent.api.configuration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
