+++ Rent/Rent-API/src/main/java/com/rent/api/configuration/RootConfiguration.java
@@ -1,7 +1,6 @@
 package com.rent.api.configuration;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
@@ -9,7 +8,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
  * Created by Duck on 10/19/2016.
  */
 @Configuration
-@ComponentScan
 public class RootConfiguration {
     @Bean
     public BCryptPasswordEncoder passwordEncoder() {
