+++ Rent/Rent-API/src/main/java/com/rent/api/controllers/GeneralController.java
@@ -1,6 +1,6 @@
 package com.rent.api.controllers;
 
-import Utilities.Constants;
+import com.rent.utility.Constants;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
