+++ Rent/Rent-API/src/main/java/com/rent/api/services/listing/UserPostService.java
@@ -1,7 +1,7 @@
 package com.rent.api.services.listing;
 
 import com.rent.api.dto.listing.UserPostDTO;
-import Utilities.Filters.UserPostFilter;
+import com.rent.utility.filters.UserPostFilter;
 import com.rent.data.dataaccess.api.entities.listing.UserPost;
 
 import java.util.List;
