+++ Rent/Rent-Oauth/src/main/java/com/rent/auth/configuration/MailConfig.java
@@ -2,7 +2,6 @@ package com.rent.auth.configuration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
