+++ Rent/Rent-Oauth/src/main/java/com/rent/auth/configuration/OAuthServerConfiguration.java
@@ -1,9 +1,8 @@
 package com.rent.auth.configuration;
 
-import Utilities.Constants;
+import com.rent.utility.Constants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.AuthenticationManager;
