+++ Rent/Rent-Oauth/src/main/java/com/rent/auth/service/login/UserDetailsServiceImpl.java
@@ -4,7 +4,6 @@ import com.rent.data.dataaccess.auth.dao.user.UserDetailsDAO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
