+++ Rent/Rent-Util/src/main/java/com/rent/utility/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package Utilities;
+package com.rent.utility;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
